move rgbcolour into the material.rs
