+++ src/vec3.rs
@@ -1,8 +1,5 @@
 use rand::{self, Rng};
-use std::{
-    fmt,
-    ops::{Add, Div, Index, Mul, Neg, Sub},
-};
+use std::ops::{Add, Div, Index, Mul, Neg, Sub};
 
 pub type Point = Vec3;
 
@@ -173,31 +170,5 @@ impl Vec3 {
         let r_out_perp = etai_over_etat * (self + cos_theta * normal);
         let r_out_parallel = -(1.0 - r_out_perp.length_squared()).abs().sqrt() * normal;
         r_out_perp + r_out_parallel
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Default)]
-pub struct RGBColour {
-    r: u8,
-    g: u8,
-    b: u8,
-}
-impl fmt::Display for RGBColour {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} {} {}", self.r, self.g, self.b)
-    }
-}
-impl From<Vec3> for RGBColour {
-    fn from(other: Vec3) -> Self {
-        Self {
-            g: (other.e[1].sqrt() * 255.999) as u8,
-            b: (other.e[2].sqrt() * 255.999) as u8,
-            r: (other.e[0].sqrt() * 255.999) as u8,
-        }
-    }
-}
-impl From<&RGBColour> for [u8; 3] {
-    fn from(colour: &RGBColour) -> Self {
-        [colour.r, colour.g, colour.b]
     }
 }
